Separated commands instructions from main() and created new source command.rs
