+++ src/commands.rs
@@ -1,0 +1,230 @@
+use crate::Cli;
+use crate::Commands;
+use chrono::NaiveTime;
+use r_timelog::config::Config;
+use r_timelog::{db, logic};
+use rusqlite::Connection;
+
+/// Handle the `init` command
+pub fn handle_init(cli: &Cli, db_path: &str) -> rusqlite::Result<()> {
+    if let Some(custom) = &cli.db {
+        Config::init_all(Some(custom.clone()), cli.test).unwrap();
+    } else {
+        Config::init_all(None, cli.test).unwrap();
+    }
+
+    if cli.test {
+        // In test mode, use db_path directly
+        let conn = Connection::open(db_path)?;
+        db::init_db(&conn)?;
+        println!("✅ Test database initialized at {}", db_path);
+    } else {
+        // Production mode: reload config
+        let config = Config::load();
+        let conn = Connection::open(&config.database)?;
+        db::init_db(&conn)?;
+        println!("✅ Database initialized at {}", config.database);
+    }
+
+    Ok(())
+}
+
+/// Handle the `add` command
+pub fn handle_add(cmd: &Commands, db_path: &str) -> rusqlite::Result<()> {
+    if let Commands::Add {
+        date,
+        pos_pos,
+        start_pos,
+        lunch_pos,
+        end_pos,
+        pos,
+        start,
+        lunch,
+        end,
+    } = cmd
+    {
+        let conn = Connection::open(db_path)?;
+
+        // validate date
+        if chrono::NaiveDate::parse_from_str(date, "%Y-%m-%d").is_err() {
+            eprintln!("❌ Invalid date format: {} (expected YYYY-MM-DD)", date);
+            return Ok(());
+        }
+
+        // merge positional and option values
+        let pos = pos.clone().or(pos_pos.clone());
+        let start = start.clone().or(start_pos.clone());
+        let lunch = (*lunch).or(*lunch_pos);
+        let end = end.clone().or(end_pos.clone());
+
+        // Handle position
+        if let Some(p) = pos.as_ref() {
+            let p = p.trim().to_uppercase();
+            if p != "O" && p != "R" {
+                eprintln!("❌ Invalid position: {} (use O=office or R=remote)", p);
+                return Ok(());
+            }
+            db::upsert_position(&conn, date, &p)?;
+            println!("✅ Position {} set for {}", p, date);
+        }
+
+        // Handle start time
+        if let Some(s) = start.as_ref() {
+            if NaiveTime::parse_from_str(s, "%H:%M").is_err() {
+                eprintln!("❌ Invalid start time: {} (expected HH:MM)", s);
+                return Ok(());
+            }
+            db::upsert_start(&conn, date, s)?;
+            println!("✅ Start time {} registered for {}", s, date);
+        }
+
+        // Handle lunch
+        if let Some(l) = lunch {
+            if !(0..=90).contains(&l) {
+                eprintln!(
+                    "❌ Invalid lunch break: {} (must be between 0 and 90 minutes)",
+                    l
+                );
+                return Ok(());
+            }
+            db::upsert_lunch(&conn, date, l)?;
+            println!("✅ Lunch {} min registered for {}", l, date);
+        }
+
+        // Handle end time
+        if let Some(e) = end.as_ref() {
+            if NaiveTime::parse_from_str(e, "%H:%M").is_err() {
+                eprintln!("❌ Invalid end time: {} (expected HH:MM)", e);
+                return Ok(());
+            }
+            db::upsert_end(&conn, date, e)?;
+            println!("✅ End time {} registered for {}", e, date);
+        }
+
+        // Warn if no field provided
+        if pos.is_none() && start.is_none() && lunch.is_none() && end.is_none() {
+            eprintln!("⚠️ Please provide at least one of: position, start, lunch, end");
+        }
+    }
+
+    Ok(())
+}
+
+/// Handle the `list` command
+pub fn handle_list(period: Option<String>, db_path: &str) -> rusqlite::Result<()> {
+    let conn = Connection::open(db_path)?;
+    let sessions = db::list_sessions(&conn, period.as_deref())?;
+
+    if sessions.is_empty() {
+        println!("⚠️  No recorded sessions found");
+        return Ok(());
+    }
+
+    if let Some(p) = period {
+        if p.len() == 4 {
+            println!("📅 Saved sessions for year {}:", p);
+        } else if p.len() == 7 {
+            let parts: Vec<&str> = p.split('-').collect();
+            let year = parts[0];
+            let month = parts[1];
+            println!(
+                "📅 Saved sessions for {} {}:",
+                logic::month_name(month),
+                year
+            );
+        }
+    } else {
+        println!("📅 Saved sessions:");
+    }
+
+    for s in sessions {
+        let has_start = !s.start.trim().is_empty();
+        let has_end = !s.end.trim().is_empty();
+
+        if has_start && !has_end {
+            // Only start → calculate expected end
+            let expected = logic::calculate_expected_exit(&s.start, 0);
+            println!(
+                "{:>3}: {} | Position {} | Start {} | \x1b[90mLunch   -\x1b[0m    | \x1b[90mEnd   -\x1b[0m   | Expected {} | \x1b[90mSurplus   -\x1b[0m",
+                s.id,
+                s.date,
+                s.position,
+                s.start,
+                expected.format("%H:%M"),
+            );
+        } else if has_start && has_end {
+            let start_time = NaiveTime::parse_from_str(&s.start, "%H:%M").unwrap();
+            let end_time = NaiveTime::parse_from_str(&s.end, "%H:%M").unwrap();
+            let pos_char = s.position.chars().next().unwrap_or('O');
+            let crosses_lunch = logic::crosses_lunch_window(&s.start, &s.end);
+
+            // Compute effective lunch
+            let effective_lunch =
+                logic::effective_lunch_minutes(s.lunch, &s.start, &s.end, pos_char);
+
+            if crosses_lunch && effective_lunch > 0 {
+                // Case with lunch (inserted or automatic)
+                let expected = logic::calculate_expected_exit(&s.start, effective_lunch);
+                let surplus = logic::calculate_surplus(&s.start, effective_lunch, &s.end);
+                let surplus_minutes = surplus.num_minutes();
+
+                let color_code = if surplus_minutes < 0 {
+                    "\x1b[31m"
+                } else if surplus_minutes > 0 {
+                    "\x1b[32m"
+                } else {
+                    "\x1b[0m"
+                };
+
+                let formatted_surplus = if surplus_minutes == 0 {
+                    "0".to_string()
+                } else {
+                    format!("{:+}", surplus_minutes)
+                };
+
+                println!(
+                    "{:>3}: {} | Position {} | Start {} | Lunch {:02} min | End {} | Expected {} | Surplus {}{:>4} min\x1b[0m",
+                    s.id,
+                    s.date,
+                    s.position,
+                    s.start,
+                    effective_lunch,
+                    s.end,
+                    expected.format("%H:%M"),
+                    color_code,
+                    formatted_surplus
+                );
+            } else {
+                // Case without lunch (work entirely outside the window)
+                let duration = end_time - start_time;
+                println!(
+                    "{:>3}: {} | Position {} | Start {} | \x1b[90mLunch   -\x1b[0m    | End {} | \x1b[36mWorked {:>2} h {:02} min\x1b[0m",
+                    s.id,
+                    s.date,
+                    s.position,
+                    s.start,
+                    s.end,
+                    duration.num_hours(),
+                    duration.num_minutes() % 60
+                );
+            }
+        } else {
+            // Incomplete information
+            println!(
+                "{:>3}: {} | Position {} | Start {} | Lunch {} | End {}",
+                s.id,
+                s.date,
+                s.position,
+                if has_start { &s.start } else { "-" },
+                if s.lunch > 0 {
+                    format!("{} min", s.lunch)
+                } else {
+                    "-".to_string()
+                },
+                if has_end { &s.end } else { "-" }
+            );
+        }
+    }
+
+    Ok(())
+}
